requirements_full.txt
# Core Dependencies
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0

# Web Framework
Flask>=2.3.0
flask-cors>=4.0.0

# Authentication & Security
Flask-JWT-Extended>=4.5.0
bcrypt>=4.0.0

# Database
pymongo>=4.5.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
Pillow>=10.0.0
